+++ CHANGELOG.md
@@ -26,12 +26,18 @@ https://github.com/nwnxee/unified/compare/build8193.34...HEAD
 - Area: GetTileInfoByTileIndex()
 - Area: GetPathExists()
 - Area: {Get|Set}AreaFlags()
+- Area: GetAreaWind()
 - Creature: {Get|Set}SkillPointsRemainingByLevel()
 - Creature: {Get|Set}InitiativeModifier()
+- Creature: GetBodyBag()
+- Creature: GetMovementRateFactorCap()
 - Effect: AccessorizeVisualEffect()
+- Encounter: Destroy()
 - Events: SubscribeEventScriptChunk()
 - Events: UnsubscribeEventScriptChunk()
 - Events: GetNumSubscribers()
+- Object: SetConversationPrivate()
+- Player: UpdateWind();
 - Regex: Match()
 
 ### Changed